Create reeborg hurdle 4 most extensive while if combination lesson
Most exhaustive learnign curve, covering all elements of while if combinations. Practice this once again. To break a while loop simple include the next condition in the same line of indentation where 'while' begins - outside the while loop, that way the condition is executed when while is not TRUE.  

Great learning curve.